removed warnings related to change in getNumPoints() return type.
git-svn-id: http://svn.osgeo.org/geos/trunk@1494 5242fede-7e19-0410-aef8-94bd7d2200fb

+++ source/headers/geos/operation/predicate/RectangleIntersects.h
@@ -57,7 +57,7 @@ public:
 	 * Must be determined empirically.  Should err on the
 	 * safe side by making value smaller rather than larger.
 	 */
-	static const int MAXIMUM_SCAN_SEGMENT_COUNT;
+	static const size_t MAXIMUM_SCAN_SEGMENT_COUNT;
 
 	/** \brief
 	 * Create a new intersects computer for a rectangle.
